Code refactor and clutter + Update readme
